+++ wetterdienst/dwd/observations/__init__.py
@@ -2,7 +2,6 @@
 # Copyright (c) 2018-2021, earthobservations developers.
 # Distributed under the MIT License. See LICENSE for more info.
 from wetterdienst.dwd.observations.api import (
-    DwdObservationMetadata,
     DwdObservationRequest,
     DwdObservationValues,
 )