+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Development
 - Use direct mapping to get a parameter set for a parameter
 - Rename DwdObservationParameterSet to DwdObservationDataset as well as corresponding
   columns
+- Merge metadata access into Request
 
 0.15.0 (07.03.2021)
 *******************